requirements.txt
beautifulsoup4==4.6.0
boto3==1.7.26
botocore==1.10.26
bs4==0.0.1
certifi==2018.4.16
chardet==3.0.4
cssselect==1.0.3
docutils==0.14
fake-useragent==0.1.10
future==0.16.0
idna==2.6
jmespath==0.9.3
oauthlib==2.1.0
parse==1.8.4
pyboto3==1.4.4
pyee==5.0.0
pyppeteer==0.0.17
pyquery==1.4.0
python-dateutil==2.7.3
python-twitter==3.4.1
requests==2.18.4
requests-html==0.9.0
requests-oauthlib==0.8.0
s3transfer==0.1.13
six==1.11.0
urllib3==1.22
w3lib==1.19.0
websockets==5.0.1
